requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements/requirements.txt requirements/requirements.in
#
asgiref==3.6.0
    # via django
django==4.1.7
    # via
    #   -r requirements/requirements.in
    #   djangorestframework
djangorestframework==3.14.0
    # via -r requirements/requirements.in
djangorestframework-camel-case==1.4.2
    # via -r requirements/requirements.in
pytz==2022.7.1
    # via djangorestframework
sqlparse==0.4.3
    # via django
